Reorganized the PInvoke classes
The PInvoke classes are now organized by DLL they are imported from.
They are also in alphabetic order. It should stay this way! Makes it way
easier to find stuff. Had to do some fix up of scripts to accomdate this
change. There are also sections for enums, structs and helper classes.